+++ frontend/src/stores/session.ts
@@ -45,7 +45,7 @@ export const sessionStore = defineStore('mail-session', () => {
 	})
 
 	const branding = createResource({
-		url: 'mail.api.mail.get_branding',
+		url: 'mail.api.get_branding',
 		cache: 'brand',
 		auto: true,
 		onSuccess(data) {
