+++ mail/api/__init__.py
@@ -1,0 +1,25 @@
+import frappe
+from frappe.translate import get_all_translations
+
+
+@frappe.whitelist(allow_guest=True)
+def get_branding() -> dict:
+	"""Returns branding information."""
+
+	return {
+		"brand_name": frappe.db.get_single_value("Website Settings", "app_name"),
+		"brand_html": frappe.db.get_single_value("Website Settings", "brand_html"),
+		"favicon": frappe.db.get_single_value("Website Settings", "favicon"),
+	}
+
+
+@frappe.whitelist(allow_guest=True)
+def get_translations() -> dict:
+	"""Returns translations for the current user's language."""
+
+	if frappe.session.user != "Guest":
+		language = frappe.db.get_value("User", frappe.session.user, "language")
+	else:
+		language = frappe.db.get_single_value("System Settings", "language")
+
+	return get_all_translations(language)
